+++ src/dbus_fast/_private/marshaller.py
@@ -1,4 +1,4 @@
-from struct import Struct, error, pack
+from struct import Struct, error
 from typing import Any, Callable, Dict, Iterable, List, Optional, Tuple
 
 from ..signature import SignatureTree, SignatureType, Variant
@@ -7,39 +7,48 @@ PACK_UINT32 = Struct("<I").pack
 
 
 class Marshaller:
-    def __init__(self, signature: str, body: Any) -> None:
+    """Marshall data for Dbus."""
+
+    __slots__ = ("signature_tree", "_buf", "body")
+
+    def __init__(self, signature: str, body: List[Any]) -> None:
+        """Marshaller constructor."""
         self.signature_tree = SignatureTree._get(signature)
-        self.buffer = bytearray()
+        self._buf = bytearray()
         self.body = body
 
+    @property
+    def buffer(self):
+        return self._buf
+
     def align(self, n) -> int:
-        offset = n - len(self.buffer) % n
+        offset = n - len(self._buf) % n
         if offset == 0 or offset == n:
             return 0
-        self.buffer.extend(bytes(offset))
+        self._buf.extend(bytes(offset))
         return offset
 
     def write_boolean(self, boolean: bool, _=None) -> int:
         written = self.align(4)
-        self.buffer.extend(PACK_UINT32(int(boolean)))
+        self._buf.extend(PACK_UINT32(int(boolean)))
         return written + 4
 
     def write_signature(self, signature: str, _=None) -> int:
-        signature = signature.encode()
+        signature_bytes = signature.encode()
         signature_len = len(signature)
-        self.buffer.append(signature_len)
-        self.buffer.extend(signature)
-        self.buffer.append(0)
+        self._buf.append(signature_len)
+        self._buf.extend(signature_bytes)
+        self._buf.append(0)
         return signature_len + 2
 
-    def write_string(self, value: str, _=None) -> int:
-        value = value.encode()
+    def write_string(self, value, _=None) -> int:
+        value_bytes = value.encode()
         value_len = len(value)
         written = self.align(4) + 4
-        self.buffer.extend(PACK_UINT32(value_len))
-        self.buffer.extend(value)
+        self._buf.extend(PACK_UINT32(value_len))
+        self._buf.extend(value_bytes)
         written += value_len
-        self.buffer.append(0)
+        self._buf.append(0)
         written += 1
         return written
 
@@ -52,9 +61,9 @@ class Marshaller:
         # TODO max array size is 64MiB (67108864 bytes)
         written = self.align(4)
         # length placeholder
-        offset = len(self.buffer)
+        offset = len(self._buf)
         written += self.align(4) + 4
-        self.buffer.extend(PACK_UINT32(0))
+        self._buf.extend(PACK_UINT32(0))
         child_type = type_.children[0]
 
         if child_type.token in "xtd{(":
@@ -67,13 +76,13 @@ class Marshaller:
                 array_len += self.write_dict_entry([key, value], child_type)
         elif child_type.token == "y":
             array_len = len(array)
-            self.buffer.extend(array)
+            self._buf.extend(array)
         elif child_type.token in self._writers:
             writer, packer, size = self._writers[child_type.token]
             if not writer:
                 for value in array:
                     array_len += self.align(size) + size
-                    self.buffer.extend(packer(value))
+                    self._buf.extend(packer(value))
             else:
                 for value in array:
                     array_len += writer(self, value, child_type)
@@ -84,7 +93,7 @@ class Marshaller:
 
         array_len_packed = PACK_UINT32(array_len)
         for i in range(offset, offset + 4):
-            self.buffer[i] = array_len_packed[i - offset]
+            self._buf[i] = array_len_packed[i - offset]
 
         return written + array_len
 
@@ -109,7 +118,7 @@ class Marshaller:
         writer, packer, size = self._writers[t]
         if not writer:
             written = self.align(size)
-            self.buffer.extend(packer(body))
+            self._buf.extend(packer(body))
             return written + size
         return writer(self, body, type_)
 
@@ -119,10 +128,10 @@ class Marshaller:
             self._construct_buffer()
         except error:
             self.signature_tree.verify(self.body)
-        return self.buffer
+        return self._buf
 
     def _construct_buffer(self):
-        self.buffer.clear()
+        self._buf.clear()
         for i, type_ in enumerate(self.signature_tree.types):
             t = type_.token
             if t not in self._writers:
@@ -132,11 +141,11 @@ class Marshaller:
             if not writer:
 
                 # In-line align
-                offset = size - len(self.buffer) % size
+                offset = size - len(self._buf) % size
                 if offset != 0 and offset != size:
-                    self.buffer.extend(bytes(offset))
+                    self._buf.extend(bytes(offset))
 
-                self.buffer.extend(packer(self.body[i]))
+                self._buf.extend(packer(self.body[i]))
             else:
                 writer(self, self.body[i], type_)